+++ Docs/Whats_New.md
@@ -16,7 +16,7 @@ For best results, use a Markdown reader.*
 MoltenVK 1.0.31
 ---------------
 
-Released 2019/01/15
+Released 2019/01/16
 
 - Support runtime config via runtime environment variables
 - Add full ImageView swizzling to config, and disable it by default.
@@ -31,6 +31,7 @@ Released 2019/01/15
 - Update copyright to 2019.
 - Advertise the `VK_AMD_gpu_shader_half_float` extension.
 - Support the `VK_KHR_variable_pointers` extension.
+- MoltenVKShaderConverter tool exit with fail code on any file conversion fail.
 - Update to latest SPIRV-Cross version:
 	- MSL: Support SPV_KHR_variable_pointers.
 	- MSL: Workaround missing gradient2d() on macOS for typical cascaded shadow mapping.